Changed Tab from using constraint based layout to frame based layout, for its active and inactive images. So subclasses are free to set their own constraints if they want, and free to also animate those views by frame, besides constraints.
Fixed a bug where setImages() will crash in subclass overrides, due to it being called in the base init, at which time, those chrome elements will not exist yet.

Fixed a bug where subclasses that set their own positioning of the active and inactive images in place of the default, would have those changes overwritten if a size changing layout style is used. Because those layout styles, after changing the size, would then 'self-correct' the positioning back to the default.

Created a new method layoutContent() that sets the default positioning of the active and inactive images. Subclasses can then override this method to specify their own positioning. Or to layout their chrome elements by frame (versus by constraints in the init), without having anything overwritten, due to the ordering of calls that the tab bar makes when laying out and readying its tabs.